LLMChat-v2.0.0/requirements.txt
# Core dependencies for Voyeur Chat
pyperclip>=1.8.2
faiss-cpu>=1.7.4
numpy==1.26.4
PyPDF2>=3.0.1
aiosqlite>=0.21.0
requests>=2.32.3
pydub>=0.25.1
simpleaudio>=1.0.4
google-cloud-texttospeech>=2.14.2
google-cloud-speech>=2.27.0
pyttsx3>=2.98
whisperx>=3.3.1
pyannote-audio==3.3.2
pyaudio>=0.2.14

# Dependencies for Sesame CSM (csm)
torch==2.2.2 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.2.2 --index-url https://download.pytorch.org/whl/cpu
transformers==4.44.0
huggingface_hub==0.28.1
soundfile>=0.13.1
ffmpeg-python>=0.2.0
torchtune>=0.3.0
torchao==0.0.3
./triton/python  # Local triton repository (correct subdirectory)
moshi==0.2.4
./silentcipher  # Local modified silentcipher

# Additional dependencies for compatibility
litellm>=1.56.5
openai>=1.78.0
tenacity>=9.1.0